+++ Source/ModuleLightEva.cs
@@ -60,6 +60,27 @@ public class ModuleLightEva : ModuleLight {
   [KSPField(guiName = "#SurfaceLights_ModuleLightEva_LightRange", advancedTweakable = true)]
   [UI_FloatRange(stepIncrement = 0.1f, minValue = 0f, maxValue = 100f)]
   public float lightRange;
+
+  /// <summary>Resets all the adjustable settings to the state on the scene load.</summary>
+  /// <summary>Once the scene is saved, this state becomes the base.</summary>
+  [KSPEvent(guiName = "#SurfaceLights_ModuleLightEva_ResetUnsaved", guiActiveEditor = true)]
+  public void ResetUnsavedSettings() {
+    RestorePersistedSettings();
+  }
+  #endregion
+
+  #region Local fields and properties
+  /// <summary>Spot angle of the light at the scene load.</summary>
+  /// <seealso cref="SavePersistedSettings"/>
+  float _originalSpotAngle;
+
+  /// <summary>Light's range at the scene load.</summary>
+  /// <seealso cref="SavePersistedSettings"/>
+  float _originalLightRange;
+
+  /// <summary>Light's color at the scene load.</summary>
+  /// <seealso cref="SavePersistedSettings"/>
+  Color _originalColor;
   #endregion
 
   #region PartModule overrides
@@ -75,6 +96,7 @@ public class ModuleLightEva : ModuleLight {
     if (!allowEvaControl) {
       return;  // Nothing to do.
     }
+    SavePersistedSettings();
 
     // Populate the UI controls with the actual data from the model.
     var refLight = lights[0];
@@ -125,6 +147,34 @@ public class ModuleLightEva : ModuleLight {
   protected static void SetupFieldForEva(BaseField kspField) {
     kspField.guiActiveUnfocused = true;
     kspField.guiActive = true;
+  }
+
+  /// <summary>Saves the light settings from the loaded part state.</summary>
+  /// <remarks>
+  /// This state is already affected by any overrides. It's not the default settings in the part
+  /// prefab.
+  /// </remarks>
+  /// <seealso cref="RestorePersistedSettings"/>
+  protected virtual void SavePersistedSettings() {
+    var refLight = lights[0];
+    _originalSpotAngle = refLight.spotAngle;
+    _originalLightRange = refLight.range;
+    _originalColor = new Color(lightR, lightG, lightB);
+  }
+
+  /// <summary>Restores light settings to what was in effect at the part load.</summary>
+  /// <remarks>
+  /// This method must restore values via the <c>BasicField.SetValue</c> method. Adjusting the part
+  /// in any other way would likely introduce a state inconsistency. 
+  /// </remarks>
+  /// <seealso cref="SavePersistedSettings"/>
+  /// <seealso cref="ResetUnsavedSettings"/>
+  protected virtual void RestorePersistedSettings() {
+    SetupField(nameof(spotAngle), f => f.SetValue(_originalSpotAngle, this));
+    SetupField(nameof(lightRange), f => f.SetValue(_originalLightRange, this));
+    SetupField(nameof(lightR), f => f.SetValue(_originalColor.r, this));
+    SetupField(nameof(lightG), f => f.SetValue(_originalColor.g, this));
+    SetupField(nameof(lightB), f => f.SetValue(_originalColor.b, this));
   }
 
   /// <summary>Applies a setup function on a KSP part module event.</summary>